Moved some deadtime related funcitons into deadtime class and added events
